# perf: write utf8 bytes directly into destination.