**Change:**
Encoding.UTF8.GetBytes allows writing bytes into the destination with offset directly, where is no need for intermediary buffer allocation and extra copy operation.

**Performance benchmarks:**
```
BenchmarkDotNet v0.13.12, macOS Sonoma 14.4.1 (23E224) [Darwin 23.4.0]
Intel Core i7-1068NG7 CPU 2.30GHz, 1 CPU, 8 logical and 4 physical cores
.NET SDK 8.0.100
  [Host]     : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2
  DefaultJob : .NET 8.0.0 (8.0.23.53103), X64 RyuJIT AVX2


| Method | Mean      | Error     | StdDev    | Gen0   | Allocated |
|------- |----------:|----------:|----------:|-------:|----------:|
| OldWay | 21.933 ns | 0.4508 ns | 0.7406 ns | 0.0076 |      32 B |
| NewWay |  8.698 ns | 0.0364 ns | 0.0304 ns |      - |         - |
```
Value used for testing -> "test1"