@shannonklaus 2024-07-25
Thank you for your pull request. It will be included in my next release, which is happening by the end of the week.